Refactor to get all cities with async thunk
